Update transition relation to work on VSS
The transition relation should work on VSS to reflect more closely
the automaton displayed in section 5 of hybrid level 3 specification.

As a result, the event handling is done in a separate procedure
Handle_Event, which is responsible for updating the state of the TDD
and call Transition_Relation on relevant VSS (it could call it on all VSS,
but there is no need to).

We also separate the static track data (relationship between VSS and TDD)
from the non-static track data (states), to simplify the work of provers,
and as a result simplify also contracts & ghost code. Indeed, the static
track data never changes, so better to have it as a separate variable
Track_Static that is not modified by the program.